cpptooling: improve search for compilation flags
A fallback strategy is added which analyzes the files in the compilation
database by parsing them and looking for include directives.

This improves quality of life for users when they want to create a test
double for e.g. a c++ interface.

+++ source/cpptooling/analyzer/clang/cursor_visitor.d
@@ -23,7 +23,7 @@ private @safe nothrow struct ASTCursor {
 private nothrow struct AST_BreathFirstResult {
     import std.container : Array;
 
-    private size_t depth_;
+    private int depth_;
     private typeof(Array!(Cursor).opSlice()) r;
     // index 0: the current range that is operated on.
     // index 1: the next one that is being filled with data.
@@ -65,7 +65,7 @@ private nothrow struct AST_BreathFirstResult {
         return r.empty && data[1].empty;
     }
 
-    size_t depth() {
+    int depth() {
         return depth_;
     }
 }
